requirements.txt
openai
PyMuPDF
langchain-community
HuggingFaceEmbeddings
Faiss
langchain_openai
pydantic